바로 결제 분기처리 완료 , 배열에 null 이 담겨있으면 값이 있음으로 처리되기때문에 빈배열 null 값체크 곶건 필요

+++ src/main/java/com/example/team01/payment/PaymentController.java
@@ -23,6 +23,7 @@ import org.springframework.web.bind.annotation.*;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Slf4j
@@ -76,7 +77,7 @@ public class PaymentController {
     @PostMapping()
     public ResponseEntity<?> postPayment(@RequestBody PaymentVO paymentVO,@AuthenticationPrincipal PrincipalDetails userDetails) {
         // 결제한 데이터 결제 테이블에 넣어주기
-        log.info("postPayment ------------------paymentVO:{}", paymentVO);
+
         // 1. Null 체크
         if (paymentVO == null) {
             return ResponseEntity.badRequest().body("결제 정보가 없습니다.");
@@ -103,11 +104,24 @@ public class PaymentController {
         paymentService.insertPayment(paymentVO,clientId);
         //payment_cart 결제내역테이블에 데이터 insert ( 목적: 도서정보조회 , 결제상세내역 조회 )
         paymentService.insertPaymentList(paymentVO);
+
+        log.info("paymentVO-------------:{}",paymentVO);
+        log.info("paymentVO.getCartIds()-------------:{}",paymentVO.getCartIds());
+
         // 장바구니 목록에서 delete //cartIds=[26, 24, 25] 파라미터를 List로 전달 ==> IN 절 다중 조회(List 파라미터 IN 절) 사용해서 삭제
-        if (paymentVO.getCartIds() != null && !paymentVO.getCartIds().isEmpty()) { //  바로구매의 경우 cartIds가 없기때문에 조건분기
-            log.info("단건구매결제 삭제 왜 실행안돼냐아----- 장바구니 아이디가 없어서 여기에 안걸림:{}",paymentVO.getCartIds());
+        if (paymentVO.getCartIds() != null
+                && !paymentVO.getCartIds().isEmpty()
+                && paymentVO.getCartIds().stream().anyMatch(Objects::nonNull)
+            // cartIds=[null] => null 이 담긴 빈 배열은 true 처리(값이 있다고 판단)되어서 조건 추가 필요
+        ) {
+            log.info("단건구매 결제, 장바구니 결제는 여기에서 처리 paymentVO.getCartIds(): {}",paymentVO.getCartIds());
             cartService.deleteToCartList(paymentVO.getCartIds());
+        }else{ // 바로구매 조건분기 ( cartId가 null 인 경우 처리 )
+            log.info("바로구매결제 실행 처리 분기 ----- 장바구니 삭제 생략:{}",paymentVO.getCartIds());
+            return ResponseEntity.ok(Map.of("message", "결제 성공", "payId", paymentVO.getPayId()));
+
         }
+
         // 결제 완료 후 payId 넘겨주기 ==> SPA 경우에는 프론트에서 이동을 제어하는게 더 나음, payId를 공통으로 가지고 잇으니까 같이 넘겨줌
         return ResponseEntity.ok(Map.of("message", "결제 성공", "payId", paymentVO.getPayId()));
     }
